Updated README and added styling. refactored

+++ app/controllers/application_controller.rb
@@ -101,8 +101,13 @@ class ApplicationController < Sinatra::Base
   #creates new Vacation instance out of recieved params
   #sends instance data to /vacations/show
   post '/vacations/show/:id' do
-    @vacation = Vacation.create(title: params[:title], location: params[:location], date: params[:date], description: params[:description], user_id: session[:user_id])
-    erb :'/vacations/show'
+    if params[:title].empty? || params[:description].empty?
+      @error = "Please fill in a title and description"
+      erb :'/vacations/new'
+    else
+      @vacation = Vacation.create(title: params[:title], location: params[:location], date: params[:date], description: params[:description], user_id: session[:user_id])
+      erb :'/vacations/show'
+    end
   end
 
   get '/vacations/:id/edit' do
@@ -115,16 +120,26 @@ class ApplicationController < Sinatra::Base
   end
 
   patch '/vacations/show/:id' do
-    @vacation = Vacation.find(params[:id])
-    @vacation.update(title: params[:title], location: params[:location], date: params[:date], description: params[:description])
-    erb :'/vacations/show'
+    if params[:title].empty? || params[:description].empty?
+      @vacation = Vacation.find(params[:id])
+      @error = "Please fill in a title and description"
+      erb :'/vacations/edit'
+    else
+      @vacation = Vacation.find(params[:id])
+      @vacation.update(title: params[:title], location: params[:location], date: params[:date], description: params[:description])
+      erb :'/vacations/show'
+    end
   end
 
   #Destroy->send delete request to vacations/show with params
   delete '/vacations/show/:id' do
-    vacation = Vacation.find(params[:id])
-    vacation.destroy
-    redirect '/user/home'
+    if logged_in?
+      vacation = Vacation.find(params[:id])
+      vacation.destroy
+      redirect '/user/home'
+    else 
+      redirect '/sessions/login'
+    end
   end
 
   def logged_in?
